chore(button): move button to different directory
