+++ src/app/components/Button/Button.test.tsx
@@ -1,0 +1,16 @@
+import { shallow } from "enzyme";
+import * as React from "react";
+import { Button } from "./Button";
+
+describe("Button", () => {
+  it("should match snapshot", () => {
+    expect(shallow(<Button>Default</Button>)).toMatchSnapshot();
+    expect(shallow(<Button type={"error"}>Error</Button>)).toMatchSnapshot();
+    expect(shallow(<Button type={"primary"}>Primary</Button>)).toMatchSnapshot();
+    expect(shallow(<Button type={"secondary"}>Secondary</Button>)).toMatchSnapshot();
+    expect(shallow(<Button type={"success"}>Success</Button>)).toMatchSnapshot();
+    expect(shallow(<Button btnSize={"default"}>Success</Button>)).toMatchSnapshot();
+    expect(shallow(<Button btnSize={"small"}>Success</Button>)).toMatchSnapshot();
+    expect(shallow(<Button btnSize={"large"}>Success</Button>)).toMatchSnapshot();
+  });
+});
