+++ src/app/components/Button/Button.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { classes, stylesheet } from "typestyle";
-import { Color } from "../constants/Color";
+import { Color } from "../../constants/Color";
 
 const classNames = stylesheet({
   button: {
